requirements.txt
appdirs==1.4.4
attrs==19.3.0
backcall==0.2.0
black==19.10b0
click==7.1.2
decorator==4.4.2
Django==2.2.15
djangorestframework==3.11.1
flake8==3.8.3
flake8-colors==0.1.6
iniconfig==1.0.1
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.2.5
